requirements.txt
asgiref==3.6.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
dill==0.3.6
Django==4.2
django-cors-headers==3.14.0
django-rest-knox==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
dol==0.1.133
front==0.1.87
i2==0.0.182
idna==3.4
meshed==0.1.110
mysqlclient==2.1.1
oauthlib==3.2.2
Pillow==9.5.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pyotp==2.8.0
pypng==0.20220715.0
python3-openid==3.2.0
pytz==2023.3
qrcode==7.4.2
requests==2.28.2
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.5.0
urllib3==1.26.15
